Big Changes
Refactored quite a bit trying to get inspector elements working to generate Armatures and add them to an inventory Serializable object. Also tried my hand at doing an inventory management system, but I will need to go back to the drawing board and do some more research on how I want to implement it. Lots of tutorials, and some of them are helpful, but trying to figure out the most performant code possible.