Armor Gen and Inventory manager Mostly Complete
Finished up the Armor Generator and Inventory Manager to get the basic functionality of adding items to the player inventory and then equipping them to the bot. Will need to work on an interface on the game side that will allow us to manually select which armor we want to equip and in which slot.